+++ tests/test_models.php
@@ -195,6 +195,8 @@ class Person extends ACLModel
         ],
         'email' => [
             'type' => Model::TYPE_STRING,
+            'validate' => 'email',
+            'title' => 'Email address',
         ],
     ];